requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
flask-login
flask-socketio
# ecdsa
# paramiko
# pycrypto
